Move server protocol classes into server folder.
